ref: rename pkg to monitoring
